+++ packages/core/src/domain/configuration/index.ts
@@ -8,4 +8,4 @@ export {
 } from './configuration'
 export type { EndpointBuilder, TrackType } from './endpointBuilder'
 export { createEndpointBuilder, buildEndpointUrl } from './endpointBuilder'
-export { computeTransportConfiguration, isIntakeUrl } from './transportConfiguration'
+export { computeTransportConfiguration } from './transportConfiguration'
